Renamed wrapper as adapter
